Updated Scala language version to 2.11
For this to work Scala libraries are updated to the latest versions too.

Scalatra 2.3 uses Scalatest 2.2. Scalatra suites mixin Matchers (as
opposed to ShouldMatchers with an earlier Scalatest). As MustMatchers
now are an alternative to Matchers and a class can mixin either one but
not both, 'must' is now replaced with 'should'.